requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
streamlit
requests
